Fix for texture not loading

+++ src/render/open-gl/Shader.cs
@@ -62,7 +62,7 @@ namespace OpenGLEngine
             gl.GetAttribLocation(handle, attribName);
 
         public void SetInt(string name, int value) =>
-            gl.Uniform1(uniformLocations[name], (uint) value);
+            gl.Uniform1(uniformLocations[name],  value);
 
         public void SetFloat(string name, float value) =>
             gl.Uniform1(uniformLocations[name], value);
